refactor: cron repository
