--- server/src/interfaces/cron.interface.ts
@@ -1,20 +1,0 @@
-export const ICronRepository = 'ICronRepository';
-
-type CronBase = {
-  name: string;
-  start?: boolean;
-};
-
-export type CronCreate = CronBase & {
-  expression: string;
-  onTick: () => void;
-};
-
-export type CronUpdate = CronBase & {
-  expression?: string;
-};
-
-export interface ICronRepository {
-  create(cron: CronCreate): void;
-  update(cron: CronUpdate): void;
-}
