+++ server/test/repositories/cron.repository.mock.ts
@@ -1,4 +1,4 @@
-import { ICronRepository } from 'src/interfaces/cron.interface';
+import { ICronRepository } from 'src/types';
 import { Mocked, vitest } from 'vitest';
 
 export const newCronRepositoryMock = (): Mocked<ICronRepository> => {
